requirements.txt
pyyaml=6.0 #used to parsing config file